update fields with floating numbers in database from int to double

+++ lnhpdWebApi/Models/MedicinalIngredientContext.cs
@@ -312,8 +312,8 @@ namespace lnhpdWebApi.Models
             item.potency_constituent = reader["potency_constituent"] == DBNull.Value ? string.Empty : reader["potency_constituent"].ToString().Trim();
             item.potency_unit_of_measure = reader["potency_unit_of_measure"] == DBNull.Value ? string.Empty : reader["potency_unit_of_measure"].ToString().Trim();
             item.quantity = reader["quantity"] == DBNull.Value ? 0 : Convert.ToDouble(reader["quantity"]);
-            item.quantity_minimum = reader["quantity_minimum"] == DBNull.Value ? 0 : Convert.ToInt32(reader["quantity_minimum"]);
-            item.quantity_maximum = reader["quantity_maximum"] == DBNull.Value ? 0 : Convert.ToInt32(reader["quantity_maximum"]);
+            item.quantity_minimum = reader["quantity_minimum"] == DBNull.Value ? 0 : Convert.ToDouble(reader["quantity_minimum"]);
+            item.quantity_maximum = reader["quantity_maximum"] == DBNull.Value ? 0 : Convert.ToDouble(reader["quantity_maximum"]);
             item.quantity_unit_of_measure = reader["quantity_unit_of_measure"] == DBNull.Value ? string.Empty : reader["quantity_unit_of_measure"].ToString().Trim();
             item.ratio_numerator = reader["ratio_numerator"] == DBNull.Value ? string.Empty : reader["ratio_numerator"].ToString().Trim();
             item.ratio_denominator = reader["ratio_denominator"] == DBNull.Value ? string.Empty : reader["ratio_denominator"].ToString().Trim();
